+++ lumber.js
@@ -14,6 +14,6 @@ program
   .command('update', 'update your models\'s definition according to your database schema')
   .command('deploy', 'deploy your admin panel application to production')
   .command('user', 'show your current logged user')
-  .command('login', 'sign in to your Forest account')
+  .command('login <email>', 'sign in to your Forest account')
   .command('logout', 'sign out of your Forest account')
   .parse(process.argv);