+++ CHANGELOG.md
@@ -11,7 +11,8 @@
 - Technical - Rename environment variable SERVER_HOST to FOREST_URL.
 - Opn - Completely remove the opn NPM package.
 - Dependencies - Fix vulnaribilities by upgrading the dependencies in the package.json.
-- Command Generate - Make the appName argument mandatory from the command line args.
+- Command Generate - Make the projectName argument mandatory from the command line args.
+- Command Login - Make the email argument mandatory from the command line args.
 - Prompt - Improve the display of success/info/error command responses.
 - Wording - Improve the readability of some prompt responses.
 - Wording - Replace all existing mentions of "back office" by "admin panel" or "admin panel API".